requirements.txt
tensorflow>=2.10
opencv-python-headless
scikit-learn
Pillow
numpy
matplotlib
